Add HandlerError
This is meant to be the 'top type' for any error that occurs withing a
handler. Any 3rd party error types, ej. minijinja::Error, will need to be
converted into HandlerError and through it implement the IntoResponse
trait. That way I can use the question mark operator inside axum handlers.